pyproject.toml
[project]
name = "network-automation"
version = "0.1.0"
description = "NetBox provisioning and network automation tooling"
requires-python = ">=3.10"
dependencies = [
    "pynetbox>=7.0.0",
    "pyyaml>=6.0",
    "requests>=2.20.0",
]

[project.scripts]
provision = "netbox.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["netbox*"]
